Add userCookieRepo
